Move saveToDisk to df
