+++ src/main/scala/io/archivesunleashed/df/package.scala
@@ -1,6 +1,12 @@
 package io.archivesunleashed
 
 import org.apache.spark.sql.functions.udf
+import org.apache.spark.sql.DataFrame
+import io.archivesunleashed.matchbox._
+import java.io.ByteArrayInputStream
+import java.io.File
+import javax.imageio.{ImageIO, ImageReader}
+import java.util.Base64
 
 /**
   * UDFs for data frames.
@@ -13,4 +19,46 @@ package object df {
   val ExtractDomain = udf(io.archivesunleashed.matchbox.ExtractDomain.apply(_: String, ""))
 
   val RemovePrefixWWW = udf[String, String](_.replaceAll("^\\s*www\\.", ""))
+
+  /**
+   * Given a dataframe, serializes the images and saves to disk
+   * @param df the input dataframe
+  */
+  implicit class SaveImage(df: DataFrame) {
+    /** 
+     * @param bytesColumnName the name of the column containing the image bytes
+     * @param fileName the name of the file to save the images to (without extension)
+     * e.g. fileName = "foo" => images are saved as foo0.jpg, foo1.jpg
+    */
+    def saveToDisk(bytesColumnName: String, fileName: String) = {
+      df.select(bytesColumnName).foreach(row => {
+        try {
+          // assumes the bytes are base64 encoded already as returned by ExtractImageDetails
+          val encodedBytes: String = row.getAs(bytesColumnName);
+          val bytes = Base64.getDecoder.decode(encodedBytes);
+          val in = new ByteArrayInputStream(bytes);
+
+          val input = ImageIO.createImageInputStream(in);
+          val readers = ImageIO.getImageReaders(input);
+          if (readers.hasNext()) {
+            val reader = readers.next()
+            reader.setInput(input)
+            val image = reader.read(0)
+
+            val format = reader.getFormatName()
+            val suffix = encodedBytes.computeHash()
+            val file = new File(fileName + "-" + suffix + "." + format);
+            if (image != null) {
+              ImageIO.write(image, format, file);
+            }
+          }
+          row
+        } catch {
+          case e: Throwable => {
+            row
+          }
+        }
+      })
+    }
+  }
 }
